requirements.txt
Flask==3.0.3
mysql-connector-python==9.0.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
